+++ test/function/util.js
@@ -64,21 +64,62 @@ assert.throws(function () {math.eval('function f(x) = a * x');});
 assert.throws(function () {math.eval('a([1,1])= [4]');});
 assert.throws(function () {math.set('a', 3)});
 
-
 // test format
-assert.equal(math.format(2/7), '0.2857142857');
+assert.equal(math.format(2/7), '0.28571');
+assert.equal(math.format(0.10400), '0.104');
+assert.equal(math.format(1000), '1000');
+assert.equal(math.format(2.3e-7), '2.3e-7');
+assert.equal(math.format(2.3e-6), '2.3e-6');
+assert.equal(math.format(2.3e-5), '2.3e-5');
+assert.equal(math.format(2.3e-4), '2.3e-4');
+assert.equal(math.format(2.3e-3), '0.0023');
+assert.equal(math.format(2.3e-2), '0.023');
+assert.equal(math.format(2.3e-1), '0.23');
+assert.equal(math.format(2.3), '2.3');
+assert.equal(math.format(2.3e1), '23');
+assert.equal(math.format(2.3e2), '230');
+assert.equal(math.format(2.3e3), '2300');
+assert.equal(math.format(2.3e4), '23000');
+assert.equal(math.format(2.3e5), '2.3e5');
+assert.equal(math.format(2.3e6), '2.3e6');
+assert.equal(math.format(math.eval('1000.000')), '1000');
+assert.equal(math.format(math.eval('1000.0010')), '1000'); // rounded off at 5 digits
+assert.equal(math.format('hello'), '"hello"');
+assert.equal(math.format(math.pi), '3.1416');
+assert.equal(math.format(math.pi * 10000), '31416');
+assert.equal(math.format(math.pi / 100), '0.031416');
+assert.equal(math.format(math.e), '2.7183');
 assert.equal(math.format([[1,2],[3,4]]), '[[1, 2], [3, 4]]');
 assert.equal(math.format([[1,2/7],['hi', math.complex(2,3)]]),
-    '[[1, 0.2857142857], ["hi", 2 + 3i]]');
-
+    '[[1, 0.28571], ["hi", 2 + 3i]]');
+assert.equal(math.format(math.divide(math.complex(2,5),3)), '0.66667 + 1.6667i');
 
 assert.equal(math.format('hello, $name!', {name: 'user'}), 'hello, user!');
 assert.equal(math.format('hello, $name.first $name.last!',
     {name: {first: 'first', last: 'last'}}),
     'hello, first last!');
 
-// TODO: test format extensively
-
-// TODO: test help
 // TODO: test import
-// TODO: test typeof
+
+// test typeof
+assert.equal(math.typeof(2), 'number');
+assert.equal(math.typeof(new Number(2)), 'number');
+assert.equal(math.typeof(2 + 3), 'number');
+assert.equal(math.typeof(2 + ''), 'string');
+assert.equal(math.typeof('hello there'), 'string');
+assert.equal(math.typeof(math.format(3)), 'string');
+assert.equal(math.typeof(math.complex(2,3)), 'complex');
+assert.equal(math.typeof([1,2,3]), 'array');
+assert.equal(math.typeof(math.matrix()), 'matrix');
+assert.equal(math.typeof(math.unit('5cm')), 'unit');
+assert.equal(math.typeof(true), 'boolean');
+assert.equal(math.typeof(false), 'boolean');
+assert.equal(math.typeof(null), 'null');
+assert.equal(math.typeof(undefined), 'undefined');
+assert.equal(math.typeof(new Date()), 'date');
+assert.equal(math.typeof(function () {}), 'function');
+assert.equal(math.typeof(math.sin), 'function');
+assert.equal(math.typeof({}), 'object');
+assert.equal(math.typeof(math), 'object');
+assert.throws(function() {math.typeof(); });
+assert.throws(function() {math.typeof(1,2); });