+++ HISTORY.md
@@ -6,6 +6,9 @@ https://github.com/josdejong/mathjs
 
 - Implemented method math.eval, which uses a readonly parser to evaluate
   expressions.
+- Improved math.format, which now returns a maximum number of digits for
+  numbers (by default 5 digits, `math.format(math.pi)` returns `3.1416`).
+- Minor bug fixes.
 
 
 ## 2013-04-13, version 0.6.0